+++ tests/cfo/tests/utils/index.js
@@ -201,7 +201,7 @@ async function setupMarket({
   const MARKET_A_USDC = await Market.load(
     provider.connection,
     marketAPublicKey,
-    { commitment: "recent" },
+    { commitment: "processed" },
     DEX_PID
   );
   for (let k = 0; k < asks.length; k += 1) {
@@ -382,7 +382,8 @@ async function signTransactions({
   wallet,
   connection,
 }) {
-  const blockhash = (await connection.getRecentBlockhash("max")).blockhash;
+  const blockhash = (await connection.getRecentBlockhash("finalized"))
+    .blockhash;
   transactionsAndSigners.forEach(({ transaction, signers = [] }) => {
     transaction.recentBlockhash = blockhash;
     transaction.setSigners(
@@ -401,7 +402,7 @@ async function signTransactions({
 async function sendAndConfirmRawTransaction(
   connection,
   raw,
-  commitment = "recent"
+  commitment = "processed"
 ) {
   let tx = await connection.sendRawTransaction(raw, {
     skipPreflight: true,
@@ -429,7 +430,7 @@ async function runTradeBot(market, provider, iterations = undefined) {
   let marketClient = await Market.load(
     provider.connection,
     market,
-    { commitment: "recent" },
+    { commitment: "processed" },
     DEX_PID
   );
   const baseTokenUser1 = (